style(v0128): 코드 정렬

+++ study/studySecurity/plateerPreBack/src/main/java/org/plateer/backlms/common/dto/PageResultDTO.java
@@ -8,7 +8,7 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 @Data
-public class PageResultDTO <E>{
+public class PageResultDTO<E> {
 
     private List<E> dtoList;
 
@@ -22,23 +22,23 @@ public class PageResultDTO <E>{
 
     private List<Integer> pageNumList;
 
-    public PageResultDTO(List<E> dtoList, Pageable pageable, long total, long totalPage ){
+    public PageResultDTO(List<E> dtoList, Pageable pageable, long total, long totalPage) {
 
         this.dtoList = dtoList;
-        this.pageNum = pageable.getPageNumber() +1;
+        this.pageNum = pageable.getPageNumber() + 1;
         this.totalPageNum = totalPage;
         this.totalCount = total;
         this.pageSize = pageable.getPageSize();
 
 
-        this.start = (int) ((Math.ceil(this.pageNum / 10.0)) * 10 ) - 9;
+        this.start = (int) ((Math.ceil(this.pageNum / 10.0)) * 10) - 9;
 
-        int tempEnd = (int)(((Math.ceil(this.pageNum / 10.0)) * 10 ));
+        int tempEnd = (int) (((Math.ceil(this.pageNum / 10.0)) * 10));
 
-        this.end = tempEnd* pageSize > totalCount ?
-                (int)(Math.ceil(totalCount/(double)pageSize)) : tempEnd;
+        this.end = (long) tempEnd * pageSize > totalCount ?
+                (int) (Math.ceil(totalCount / (double) pageSize)) : tempEnd;
 
-        this.pageNumList = IntStream.rangeClosed(start,end).boxed().collect(Collectors.toList());
+        this.pageNumList = IntStream.rangeClosed(start, end).boxed().collect(Collectors.toList());
 
     }
 }
